# src/openapi.json
{
	"openapi": "3.0.0",
	"info": {
	  "title": "Podcasts API",
	  "version": "1.0.0"
	},
	"servers": [
	  {
		"url": "http://localhost:8080"
	  }
	],
	"paths": {
	  "/shows": {
		"get": {
		  "summary": "Get a list of shows",
		  "responses": {
			"200": {
			  "description": "List of shows",
			  "content": {
				"application/json": {
				  "schema": {
					"type": "array",
					"items": {
					  "$ref": "#/components/schemas/Show"
					}
				  }
				}
			  }
			}
		  }
		}
	  },
	  "/shows/{id}": {
		"get": {
		  "summary": "Get a show by id",
		  "parameters": [
			{
			  "name": "id",
			  "in": "path",
			  "required": true,
			  "description": "Show ID",
			  "schema": {
				"type": "string"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Show object",
			  "content": {
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/Show"
				  }
				}
			  }
			}
		  }
		}
	  },
	  "/episodes/{id}": {
		"get": {
		  "summary": "Get an episode by id",
		  "parameters": [
			{
			  "name": "id",
			  "in": "path",
			  "required": true,
			  "description": "Episode ID",
			  "schema": {
				"type": "string"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Episode object",
			  "content": {
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/Episode"
				  }
				}
			  }
			}
		  }
		}
	  },
	  "/episodes/{id}/summary": {
		"get": {
		  "summary": "Get a summary of an episode by ID",
		  "parameters": [
			{
			  "name": "id",
			  "in": "path",
			  "description": "The ID of the episode to get the summary for",
			  "required": true,
			  "schema": {
				"type": "string",
				"example": "12345"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Successful response with the summary text",
			  "content": {
				"text/plain": {
				  "schema": {
					"type": "string"
				  }
				}
			  }
			}
		  }
		}
	  }
	},
	"components": {
	  "schemas": {
		"Show": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string"
			},
			"title": {
			  "type": "string"
			},
			"publisher": {
			  "type": "string"
			},
			"image": {
			  "type": "string"
			},
			"thumbnail": {
			  "type": "string"
			}
		  }
		},
		"Episode": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string"
			},
			"title": {
			  "type": "string"
			},
			"description": {
			  "type": "string"
			},
			"date": {
			  "type": "string",
			  "format": "date-time"
			},
			"audio": {
			  "type": "string"
			},
			"showId": {
			  "type": "string"
			}
		  }
		}
	  }
	}
  }
  